requirements.txt
langchain
langchain-community
langchain-chroma
chromadb
sentence-transformers
streamlit
langsmith
ragas
pypdf
langchain-groq
huggingface_hub
python-dotenv
pinecone 
langchain-pinecone
ragas 
datasets